refactor: move renderer-driven requests off the main process
主进程之前代发了大量本由渲染业务驱动的网络请求（账单 / 语音转写 / 组织 /
技能浏览 / 连接器），主进程事件循环一旦卡住会连累渲染层。本次把这些请求改
由渲染进程经 oomolFetch 直发：credentials:include 让 httpOnly 会话 cookie
自动鉴权，token 不进渲染层（守 R4）；域名从 electron/domain.ts 派生（守
R2）。主进程新增 electron/net/oomol-cors.ts 的 CORS shim（onHeadersReceived
回显 Origin + 答复预检），放行 file:// / localhost 的跨站凭证请求。

connections 与 organizations 两个 IPC service 整体删除，逻辑落到
src/lib/*-client.ts；oauth 开系统浏览器与 agent 组织作用域改经薄主进程门面
chatService.openExternalUrl / setAgentOrganization。仅 deep-link 取 token
与 agent 内核 fetch（listAuthorizedServices、标题生成）仍留主进程。

质量门全绿（ts-check / lint / format / 537 测试），本地打包产物实测 file://
origin 下匿名与认证（cookie 自动附带）请求均返回 200。
