+++ src/components/AppContext.ts
@@ -1,8 +1,6 @@
 import type { AuthService } from "../../electron/auth/common.ts"
 import type { ChatService } from "../../electron/chat/common.ts"
-import type { ConnectionsService } from "../../electron/connections/common.ts"
 import type { ModelsService } from "../../electron/models/common.ts"
-import type { OrganizationsService } from "../../electron/organizations/common.ts"
 import type { SessionService } from "../../electron/session/common.ts"
 import type { SettingsService } from "../../electron/settings/common.ts"
 import type { SkillService } from "../../electron/skills/common.ts"
@@ -14,10 +12,8 @@ import * as React from "react"
 export interface AppContextValue {
   chatService: ConnectionClientService<ChatService>
   sessionService: ConnectionClientService<SessionService>
-  connectionsService: ConnectionClientService<ConnectionsService>
   skillService: ConnectionClientService<SkillService>
   modelsService: ConnectionClientService<ModelsService>
-  organizationsService: ConnectionClientService<OrganizationsService>
   settingsService: ConnectionClientService<SettingsService>
   authService: ConnectionClientService<AuthService>
   updateService: ConnectionClientService<UpdateService>
@@ -41,20 +37,12 @@ export function useSessionService(): ConnectionClientService<SessionService> {
   return useAppContext().sessionService
 }
 
-export function useConnectionsService(): ConnectionClientService<ConnectionsService> {
-  return useAppContext().connectionsService
-}
-
 export function useSkillService(): ConnectionClientService<SkillService> {
   return useAppContext().skillService
 }
 
 export function useModelsService(): ConnectionClientService<ModelsService> {
   return useAppContext().modelsService
-}
-
-export function useOrganizationsService(): ConnectionClientService<OrganizationsService> {
-  return useAppContext().organizationsService
 }
 
 export function useSettingsService(): ConnectionClientService<SettingsService> {
