+++ src/main.tsx
@@ -3,9 +3,7 @@ import { ElectronClientAdapter } from "@oomol/connection-electron-adapter/client
 import { createRoot } from "react-dom/client"
 import { AuthService } from "../electron/auth/common.ts"
 import { ChatService } from "../electron/chat/common.ts"
-import { ConnectionsService } from "../electron/connections/common.ts"
 import { ModelsService } from "../electron/models/common.ts"
-import { OrganizationsService } from "../electron/organizations/common.ts"
 import { SessionService } from "../electron/session/common.ts"
 import { SettingsService } from "../electron/settings/common.ts"
 import { SkillService } from "../electron/skills/common.ts"
@@ -28,10 +26,8 @@ client.start()
 
 const chatService = client.use(ChatService)
 const sessionService = client.use(SessionService)
-const connectionsService = client.use(ConnectionsService)
 const skillService = client.use(SkillService)
 const modelsService = client.use(ModelsService)
-const organizationsService = client.use(OrganizationsService)
 const settingsService = client.use(SettingsService)
 const authService = client.use(AuthService)
 const updateService = client.use(UpdateService)
@@ -41,10 +37,8 @@ createRoot(rootElement).render(
     value={{
       chatService,
       sessionService,
-      connectionsService,
       skillService,
       modelsService,
-      organizationsService,
       settingsService,
       authService,
       updateService,
